+++ tests/memory/enc/basic.c
@@ -13,7 +13,7 @@
 static void _set_buffer(int* buf, size_t start, size_t end)
 {
     for (size_t i = start; i < end; i++)
-        buf[i] = i;
+        buf[i] = (int)i;
 }
 
 static void _check_buffer(int* buf, size_t start, size_t end)
